Fix MySQL auth schema foreign key

+++ backend/sql/user_sql.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import BigInteger, Column, Integer, String, ForeignKey
 from core.db import Base
 
 class User(Base):
@@ -8,5 +8,5 @@ class User(Base):
     username = Column(String(50), unique=True, nullable=False, index=True)
     email = Column(String(100), unique=True, nullable=False, index=True)
     password_hash = Column(String(255), nullable=False)
-    company_id = Column(Integer, ForeignKey("company.id"), nullable=False)
+    company_id = Column(BigInteger, ForeignKey("company.id"), nullable=False)
     role = Column(String(20), default="viewer")  # ex: admin/operator/viewer
